Fixing some static library issues.
SDL2 tries to replace many MSVCRT functions so it's not dependent on Microsoft Runtimes and that conflicts with my own usage of MSVCRT. I added HAVE_LIBC to the preprocessor definitions for the static project configurations, no actual code change necessary.

Since SDL2 is designed to be a shared library any resulting executable from my static library configurations would result in that executable having exports to SDL2 calls. This isn't really problematic, but added a code change to prevent this and clean-up any target app's exports.